# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Funding Application Platform API
  version: 1.0.0
  description: |
    RESTful API for the Funding Application Submission & Assessment Platform.

    This API enables:
    - **Applicants** to submit applications with uploaded files and confirmations
    - **Coordinators** to manage funding calls, assign assessors, and view master results
    - **Assessors** to score assigned applications against defined criteria

    ## Authentication

    All endpoints except public call listings require JWT Bearer authentication.
    Service-to-service communication uses API key authentication.

    ## Rate Limiting

    - Standard endpoints: 100 requests/minute per user
    - File upload endpoints: 10 requests/minute per user
    - Export endpoints: 5 requests/minute per user

    ## Timezone Handling

    All timestamps are stored in UTC and returned in ISO 8601 format.
    Deadlines are enforced in Europe/London timezone.
  contact:
    name: Platform Support
    email: support@fundingplatform.example.com
    url: https://fundingplatform.example.com/support
  license:
    name: Proprietary
    url: https://fundingplatform.example.com/license
  x-logo:
    url: https://fundingplatform.example.com/logo.png
    altText: Funding Platform Logo
servers:
  - url: https://api.fundingplatform.example.com/api/v1
    description: Production server
  - url: https://staging-api.fundingplatform.example.com/api/v1
    description: Staging server
...
security:
  - BearerAuth: []
paths:
...
  /assessments/{assessment_id}/submit:
    parameters:
      - name: assessment_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      tags:
        - Assessments
      summary: Submit assessment
      description: |
        Submits the assessment.

        Requirements:
        - All criteria must be scored
        - Required comments must be provided
        - COI must be declared

        On submit, the assessment is locked and timestamped.
      operationId: submitAssessment
      responses:
        '200':
          description: Assessment submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assessment'
              example:
                id: "assess-001"
                assignment_id: "assign-001"
                status: "submitted"
                submitted_at: "2026-02-15T16:45:00Z"
                total_score: 83
                scores:
                  - criterion_id: "crit-001"
                    criterion_name: "Innovation"
                    score: 22
                    max_points: 25
                    comment: "Innovative approach..."
        '400':
          description: Assessment incomplete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: "ASSESSMENT_INCOMPLETE"
                  message: "Cannot submit assessment. All criteria must be scored."
                  details:
                    missing_scores:
                      - "crit-003"
                    missing_comments:
                      - "crit-001"
                    coi_declared: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Assessment already submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service communication
...
  responses:
...
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
              details: null
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "FORBIDDEN"
              message: "You do not have permission to perform this action"
              details:
                required_role: "coordinator"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
              details: null
...
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "Request validation failed"
            details:
              type: object
              nullable: true
              description: Additional error details
              additionalProperties: true

    # --------------------------------------------------------------------------
    # PAGINATION SCHEMA
    # --------------------------------------------------------------------------
...
    Assessment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        assignment_id:
          type: string
          format: uuid
        application_id:
          type: string
          format: uuid
        application_reference:
          type: string
        assessor_id:
          type: string
          format: uuid
          description: Only visible to coordinators
        status:
          type: string
          enum: [draft, submitted, returned]
        scores:
          type: array
          items:
            $ref: '#/components/schemas/Score'
        total_score:
          type: number
          description: Sum of all criterion scores
        weighted_score:
          type: number
          description: Weighted total score
        overall_comment:
          type: string
        coi_declared:
          type: boolean
        coi_has_conflict:
          type: boolean
        coi_details:
          type: string
        submitted_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Score:
      type: object
      properties:
        criterion_id:
          type: string
        criterion_name:
          type: string
        score:
          type: integer
        max_points:
          type: integer
        weight:
          type: number
        comment:
          type: string
...